feat(bearer): simplify logic
