+++ cisco_telescope/exporter_factory.py
@@ -16,7 +16,6 @@ limitations under the License.
 
 from . import options
 from . import consts
-from cisco_opentelemetry_specifications import Consts
 
 from opentelemetry.sdk.trace.export import ConsoleSpanExporter
 from opentelemetry.exporter.otlp.proto.grpc.trace_exporter import (
@@ -28,12 +27,7 @@ from opentelemetry.exporter.otlp.proto.http.trace_exporter import (
 
 
 def _set_exporter(exporter: options.ExporterOptions, opt: options.Options):
-    if opt.is_cisco_exporter:
-        return OTLPHTTPExporter(
-            endpoint=Consts.DEFAULT_COLLECTOR_ENDPOINT,
-            headers={Consts.TOKEN_HEADER_KEY: verify_token(opt.cisco_token)},
-        )
-    elif exporter.exporter_type == consts.GRPC_EXPORTER_TYPE:
+    if exporter.exporter_type == consts.GRPC_EXPORTER_TYPE:
         return OTLPGrpcExporter(
             endpoint=exporter.collector_endpoint, headers=exporter.custom_headers
         )
@@ -46,14 +40,6 @@ def _set_exporter(exporter: options.ExporterOptions, opt: options.Options):
         return ConsoleSpanExporter(service_name=opt.service_name)
     else:
         raise ValueError("Unsupported exported type")
-
-
-def verify_token(token: str) -> str:
-    auth_prefix = "Bearer "
-    if token.startswith(auth_prefix):
-        return token
-    else:
-        return auth_prefix + token
 
 
 def init_exporters(opt: options.Options):
